% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{correlationWithSpecificGene}
\alias{correlationWithSpecificGene}
\title{Find gorrelated genes with specific gene}
\usage{
correlationWithSpecificGene(tomoObj, geneID, corMethod = "pearson")
}
\arguments{
\item{tomoObj}{tomoseq object}

\item{geneID}{Target gene ID}

\item{corMethod}{Method for calculating correlation. default is "pearson".}
}
\description{
Find gorrelated genes with specific gene
}
\examples{
data(tomoObj)
findCorrelatedGenes(tomoObj)
}
