% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{doadJunker2014}
\alias{doadJunker2014}
\title{Load data of Junker2014 from cache.}
\usage{
doadJunker2014(tomoseqrCache)
}
\arguments{
\item{tomoseqrCache}{Cache of tomoseqr. You can get it using
\code{downloadJunker2014}.}
}
\value{
List of tomo-seq data in cache.
}
\description{
Load data of Junker2014 from cache.
}
