% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{DownloadJunker2014}
\alias{DownloadJunker2014}
\title{Download  part of the Tomo-seq data published by Junker et al.}
\usage{
DownloadJunker2014(verbose = FALSE)
}
\arguments{
\item{verbose}{If you want to force downloads with or without cache,
set this to TRUE.}
}
\value{
BiocFileCache object.
}
\description{
Download  part of the Tomo-seq data published by Junker et al.
}
