% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{findAxialGenes}
\alias{findAxialGenes}
\title{Find peak genes on axial}
\usage{
findAxialGenes(tomoSeqData, genes = "all")
}
\arguments{
\item{tomoSeqData}{tomo-seq data of any axis}

\item{genes}{If run for all genes}
}
\value{
A data frame consisting of gene ID, max of expression levels of the
gene, mean of expression levels calculated by excluding the maximum value and
section number showing the maximum expression level (0 means that there is
no such section).
}
\description{
Find peak genes on axial
}
\examples{
data(testx)
findAxialGenes(testx)
}
