% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{FindCorrelatedGenes}
\alias{FindCorrelatedGenes}
\title{Find gorrelated genes}
\usage{
FindCorrelatedGenes(tomoObj)
}
\arguments{
\item{tomoObj}{tomoseq object}
}
\description{
Find gorrelated genes
}
