% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{FindSpatialGenes}
\alias{FindSpatialGenes}
\title{Find peak genes on result of reconstruction}
\usage{
FindSpatialGenes(reconstructedResult)
}
\arguments{
\item{reconstructedResult}{Result of \code{Estimate3dExpressions()}}
}
\description{
Find peak genes on result of reconstruction
}
