% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{LoadJunker2014}
\alias{LoadJunker2014}
\title{Load data of Junker2014 from cache.}
\usage{
LoadJunker2014(tomoseqrCache)
}
\arguments{
\item{tomoseqrCache}{Cache of tomoseqr. You can get it using
\code{DownloadJunker2014}.}
}
\value{
List of tomo-seq data in cache.
}
\description{
Load data of Junker2014 from cache.
}
