% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{makeTomoObjSet}
\alias{makeTomoObjSet}
\title{Make set of tomoseq objects}
\usage{
makeTomoObjSet(x, y, z, mask_shape = "rectangle", species = "")
}
\arguments{
\item{x}{marginal data (x axes)}

\item{y}{marginal data (y axes)}

\item{z}{marginal data (z axes)}

\item{mask_shape}{shape of mask.}

\item{species}{scientific name}
}
\description{
Make set of tomoseq objects
}
