% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{plot1dAllExpression}
\alias{plot1dAllExpression}
\title{Plot expressions of all genes along axes}
\usage{
plot1dAllExpression(tomo_obj, axes)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{axes}{axes (1, 2 or 3)}
}
\description{
Plot expressions of all genes along axes
}
\note{
You can do the same things with
}
