% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{PlotLossFunction}
\alias{PlotLossFunction}
\title{Plot the trend of the value of the loss function.}
\usage{
PlotLossFunction(tomoObj, geneID)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID (string)}
}
\value{
NA
}
\description{
Plot the trend of the value of the loss function.
}
\examples{
data(tomoObj)
PlotLossFunction(tomoObj, "gene2")
}
