\name{testy}
\alias{testy}
\docType{data}
\title{
    A data.frame object containing a simulated Tomo-seq data
    for y-axis sections.
}
\description{
    A data.frame object containing a simulated Tomo-seq data for y-axis
    sections. The rows represent genes. The first column contains gene IDs and
    the second and subsequent columns contain gene expression levels
    in sections.
}