\name{testz}
\alias{testz}
\docType{data}
\title{
    A data.frame object containing a simulated Tomo-seq data
    for z-axis sections.
}
\description{
    A data.frame object containing a simulated Tomo-seq data for z-axis
    sections. The rows represent genes. The first column contains gene IDs and
    the second and subsequent columns contain gene expression levels
    in sections.
}