% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{toDataFrame}
\alias{toDataFrame}
\title{Convert reconstructed matrix to data.frame}
\usage{
toDataFrame(tomo_obj, gene_ID)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID}
}
\description{
Convert reconstructed matrix to data.frame
}
