% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomoseqr.R
\name{tomoSeq}
\alias{tomoSeq}
\title{tomoSeq object}
\description{
The class that contains a list of information about each genes.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{geneList}}{A list of gene ID. It's read-only.}

\item{\code{expMatOriginal}}{Original expression matrices along each axis.
It's read-only.}

\item{\code{mask}}{3D matrix, whose elements are 0 (not included in sample)
or 1 (included in sample). It's read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tomoSeq$new()}}
\item \href{#method-Estimate3dExpressions}{\code{tomoSeq$Estimate3dExpressions()}}
\item \href{#method-PlotLossFunction}{\code{tomoSeq$PlotLossFunction()}}
\item \href{#method-ToDataFrame}{\code{tomoSeq$ToDataFrame()}}
\item \href{#method-GetReconstructedResult}{\code{tomoSeq$GetReconstructedResult()}}
\item \href{#method-Animate2d}{\code{tomoSeq$Animate2d()}}
\item \href{#method-Plot1dExpression}{\code{tomoSeq$Plot1dExpression()}}
\item \href{#method-Plot1dAllExpression}{\code{tomoSeq$Plot1dAllExpression()}}
\item \href{#method-clone}{\code{tomoSeq$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make tomoSeq object. This method is not available to users.
Instead of it, \code{\link[=MakeTomoObjSet]{MakeTomoObjSet()}} is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$new(x, y, z, maskShape = "rectangle")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data.frame object containing a simulated Tomo-seq data for
x-axis sections. The rows represent genes. The first column contains gene IDs
and the second and subsequent columns contain gene expression levels
in sections.}

\item{\code{y}}{A data.frame object containing a simulated Tomo-seq data for y-axis
sections. The rows represent genes. The first column contains gene IDs and
the second and subsequent columns contain gene expression levels in sections.}

\item{\code{z}}{A data.frame object containing a simulated Tomo-seq data for z-axis
sections. The rows represent genes. The first column contains gene IDs and
the second and subsequent columns contain gene expression levels in sections.}

\item{\code{maskShape}}{Shape of mask. You can choose the value from
\code{"rectangle"}, \code{"round"} or \code{"halfround"}. The default is \code{"rectangle"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Estimate3dExpressions"></a>}}
\if{latex}{\out{\hypertarget{method-Estimate3dExpressions}{}}}
\subsection{Method \code{Estimate3dExpressions()}}{
Reconstructs 3D expression patterns of genes which are
specified. See also \code{\link[=Estimate3dExpressions]{Estimate3dExpressions()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$Estimate3dExpressions(queries = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queries}}{A vector consists of gene IDs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotLossFunction"></a>}}
\if{latex}{\out{\hypertarget{method-PlotLossFunction}{}}}
\subsection{Method \code{PlotLossFunction()}}{
Plot the transition of the value of the loss function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$PlotLossFunction(geneID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneID}}{A gene ID as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-ToDataFrame}{}}}
\subsection{Method \code{ToDataFrame()}}{
Export a result of reconstruction as data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$ToDataFrame(geneID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneID}}{A gene ID as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetReconstructedResult"></a>}}
\if{latex}{\out{\hypertarget{method-GetReconstructedResult}{}}}
\subsection{Method \code{GetReconstructedResult()}}{
Export a result of reconstruction as 3D matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$GetReconstructedResult(geneID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneID}}{A gene ID as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animate2d"></a>}}
\if{latex}{\out{\hypertarget{method-Animate2d}{}}}
\subsection{Method \code{Animate2d()}}{
Reconstruct a 3D expression pattern of a gene.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$Animate2d(
  geneID,
  target,
  xaxis,
  yaxis,
  main,
  xlab,
  ylab,
  file,
  zlim,
  interval,
  aspectRatio = c()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneID}}{A gene ID as string.}

\item{\code{target}}{A target of exportation. You can choose from "expression",
"mask" or "unite". The default value is "expression".}

\item{\code{xaxis}}{One of the axes of reconstructed matrix,
which is used the x-axis of the animation. You can choose from 1, 2 or 3.
The default value is 1.}

\item{\code{yaxis}}{One of the axes of reconstructed matrix,
which is used the y-axis of the animation. You can choose from 1, 2 or 3.
The default value is 2.}

\item{\code{main}}{The title of animation. It is NOT the file name.
The default value is same as geneID.}

\item{\code{xlab}}{A string, that is a label of x-axis. The default value is same
as \code{xaxis}.}

\item{\code{ylab}}{A string, that is a label of y-axis. The default value is same
as \code{yaxis}.}

\item{\code{file}}{A name of GIF file that is exported. The default value is
generated using geneID, xaxis and yaxis.}

\item{\code{zlim}}{Limits of expression levels that is displayed. You can specify
it as \code{c(min, max)}. The default value is automatically calculated using
the result of reconstruction.}

\item{\code{interval}}{An interval of GIF animation. The default value is 0.1.}

\item{\code{aspectRatio}}{A 2D vector of aspect ratio of animation.
You can specify the ratio as \code{c(width, height)}. If you don't specify
the value of this parameter, the ratio is calculated based on
the number of sections along each axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot1dExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Plot1dExpression}{}}}
\subsection{Method \code{Plot1dExpression()}}{
Plot expression of single gene along axis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$Plot1dExpression(geneID, axes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geneID}}{A gene ID as string}

\item{\code{axes}}{An axis of which you want to plot expression (1, 2 or 3).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot1dAllExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Plot1dAllExpression}{}}}
\subsection{Method \code{Plot1dAllExpression()}}{
Plot expressions of all genes along an axis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$Plot1dAllExpression(axes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{axes}}{An axis of which you want to plot expression (1, 2 or 3).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomoSeq$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
