% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{animate2d}
\alias{animate2d}
\title{Animate 2D expressions along one axis and generate GIF file.}
\usage{
animate2d(
  tomo_obj,
  gene_ID,
  target = "expression",
  xaxis = 1,
  yaxis = 2,
  main = gene_ID,
  xlab = xaxis,
  ylab = yaxis,
  file = paste(gene_ID, "_", target, "_", xaxis, "_", yaxis, ".gif", sep = ""),
  zlim = NA,
  interval = 0.1,
  aspect_ratio = c()
)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID (string)}

\item{target}{"expression", "mask" or "unite" (combination of expression and
mask). Default is \code{expression}.}

\item{xaxis}{Number to specify as x-axis (1, 2 or 3). Default is \code{1}.}

\item{yaxis}{Number to specify as y-axis (1, 2 or 3). Default is \code{2}.}

\item{main}{A string used for the title of the plot. Default is \code{gene_ID}.}

\item{xlab}{Label of x axis. Default is \code{xaxis}.}

\item{ylab}{Label of y axis. Default is \code{yaxis}.}

\item{file}{Path of GIF file.}

\item{zlim}{Limit of value of heatmap. If target="mask", it is ignored.}

\item{interval}{interval of GIF animation.}

\item{aspect_ratio}{A 2D vector that represents the ratio of figure. You can
specify the ratio as \code{c(width, height)}. If you don't specify the value of
this parameter, the ratio is calculated based on the number of sections
along each axis.}
}
\description{
Animate 2D expressions along one axis and generate GIF file.
}
\note{
You can do the same thing with \code{tomo_obj$animate2d(gene_ID, ...)}.
}
