% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{animate2d}
\alias{animate2d}
\title{animate 2D expressions along axes}
\usage{
animate2d(tomo_obj, gene_ID, target = "expression")
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID (string)}

\item{target}{"expression" or "mask"}
}
\description{
animate 2D expressions along axes
}
