% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{Estimate3dExpressions}
\alias{Estimate3dExpressions}
\title{Estimate 3d expressions}
\usage{
Estimate3dExpressions(
  x,
  y,
  z,
  mask,
  query,
  numIter = 100,
  normCount = "countSum",
  normMask = TRUE
)
}
\arguments{
\item{x}{A data.frame object containing a simulated Tomo-seq data
for x-axis sections. The rows represent genes. The first column
contains gene IDs and the second and subsequent columns contain
gene expression levels in sections.}

\item{y}{A data.frame object containing a simulated Tomo-seq data for y-axis
sections. The rows represent genes. The first column contains gene IDs and
the second and subsequent columns contain gene expression levels in sections.}

\item{z}{A data.frame object containing a simulated Tomo-seq data for
z-axis sections. The rows represent genes. The first column
contains gene IDs and the second and subsequent columns contain
gene expression levels in sections.}

\item{mask}{A 3D array that represents if each boxel is included to sample.
You can make a mask using \code{masker}.}

\item{query}{Vector of gene IDs}

\item{normCount}{Specifies the method to normalize
the expression amount data.}

\item{normMask}{Whether to normalize by mask or not}

\item{tomoObj}{tomoSeq object}
}
\value{
tomoSeq object
}
\description{
Estimate 3d expressions
}
