% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{estimate3dExpressions}
\alias{estimate3dExpressions}
\title{Estimate 3d expression}
\usage{
estimate3dExpressions(tomo_obj, query)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{query}{Vector of gene ID}
}
\value{
tomo_seq object
}
\description{
Estimate 3d expression
}
\note{
You can do the same things with
\code{tomo_obj$estimate3dExpressions(query)}.
}
