% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{Estimate3dExpressions}
\alias{Estimate3dExpressions}
\title{Estimate 3d expressions}
\usage{
Estimate3dExpressions(tomoObj, query)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{query}{Vector of gene IDs}
}
\value{
tomoSeq object
}
\description{
Estimate 3d expressions
}
\note{
You can do the same things with
\code{tomoObj$Estimate3dExpressions(query)}.
}
