% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{FindAxialGenes}
\alias{FindAxialGenes}
\title{Find peak genes on axial}
\usage{
FindAxialGenes(tomoSeqData, genes = "all")
}
\arguments{
\item{tomoSeqData}{tomo-seq data of any axis}

\item{genes}{If run for all genes}
}
\description{
Find peak genes on axial
}
