% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{findCorrelatedGenes}
\alias{findCorrelatedGenes}
\title{Find gorrelated genes}
\usage{
findCorrelatedGenes(tomoObj, corMethod = "pearson")
}
\arguments{
\item{tomoObj}{tomoseq object}
}
\description{
Find gorrelated genes
}
\examples{
data(tomoObj)
findCorrelatedGenes(tomoObj)
}
