% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{getReconstructedResult}
\alias{getReconstructedResult}
\title{Get reconstructed matrix}
\usage{
getReconstructedResult(tomo_obj, gene_ID)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID}
}
\description{
Get reconstructed matrix
}
\note{
You can do the same thing with
\code{tomo_obj$getReconstructedResult(gene_ID)}.
}
