% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{GetReconstructedResult}
\alias{GetReconstructedResult}
\title{Get reconstructed matrix}
\usage{
GetReconstructedResult(tomoObj, geneID)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID}
}
\description{
Get reconstructed matrix
}
\note{
You can do the same thing with
\code{tomoObj$GetReconstructedResult(geneID)}.
}
