% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageViewer.R
\name{imageViewer}
\alias{imageViewer}
\title{Output the reconstructed expression pattern as an image.}
\usage{
imageViewer(tomoObj)
}
\arguments{
\item{tomoObj}{tomoSeq object}
}
\value{
NA
}
\description{
Output the reconstructed expression pattern as an image.
}
\examples{
if (interactive()) {
    data(tomoObj)
    imageViewer(tomoObj, "gene2")
}
}
