% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageViewer.R
\name{ImageViewer}
\alias{ImageViewer}
\title{Output the reconstructed expression pattern as an image.}
\usage{
ImageViewer(tomoObj, geneID)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{geneID as string}
}
\value{
NA
}
\description{
Output the reconstructed expression pattern as an image.
}
\examples{
if (interactive()) {
    data(tomoObj)
    ImageViewer(tomoObj, "gene2")
}
}
