% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{MakeTomoObjSet}
\alias{MakeTomoObjSet}
\title{Make set of tomoseq objects}
\usage{
MakeTomoObjSet(x, y, z, maskShape = "rectangle")
}
\arguments{
\item{x}{A data.frame object containing a simulated Tomo-seq data
for x-axis sections. The rows represent genes. The first column
contains gene IDs and the second and subsequent columns contain
gene expression levels in sections.}

\item{y}{A data.frame object containing a simulated Tomo-seq data for y-axis
sections. The rows represent genes. The first column contains gene IDs and
the second and subsequent columns contain gene expression levels in sections.}

\item{z}{A data.frame object containing a simulated Tomo-seq data for
z-axis sections. The rows represent genes. The first column
contains gene IDs and the second and subsequent columns contain
gene expression levels in sections.}

\item{maskShape}{shape of mask.}
}
\description{
Make set of tomoseq objects
}
