% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{Plot1dAllExpression}
\alias{Plot1dAllExpression}
\title{Plot expressions of all genes along an axis}
\usage{
Plot1dAllExpression(tomoObj, axes)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{axes}{axis (1, 2 or 3)}
}
\description{
Plot expressions of all genes along an axis
}
\note{
You can do the same thing with
}
