% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{Plot1dAllExpression}
\alias{Plot1dAllExpression}
\title{Plot expressions of all genes along an axis}
\usage{
Plot1dAllExpression(tomoSeqData, ...)
}
\arguments{
\item{tomoSeqData}{A data.frame object containing a Tomo-seq data
for any sections. The rows represent genes. The first column
contains gene IDs and the second and subsequent columns contain
gene expression levels in sections.}

\item{...}{Arguments which are related to plot parameters.
Prease refer to \code{\link[graphics]{plot}}.}
}
\value{
NA
}
\description{
Plot expressions of all genes along an axis
}
