% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{plot1dExpression}
\alias{plot1dExpression}
\title{Plot expression of single gene along an axis}
\usage{
plot1dExpression(tomoObj, geneID, axes)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID (string)}

\item{axes}{axis ("x", "y" or "z")}
}
\value{
NA
}
\description{
Plot expression of single gene along an axis
}
\examples{
data(tomoObj)
plot1dExpression(tomoObj, "gene2", "x")
}
