% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{Plot1dExpression}
\alias{Plot1dExpression}
\title{Plot expression of single gene along an axis}
\usage{
Plot1dExpression(tomoObj, geneID, axes)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID (string)}

\item{axes}{axis ("x", "y" or "z")}
}
\description{
Plot expression of single gene along an axis
}
\note{
You can do the same thing with
\code{tomoObj$Plot1dExpression(geneID, axes)}.
}
