% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{plot1dExpression}
\alias{plot1dExpression}
\title{Plot expression of single gene along an axis}
\usage{
plot1dExpression(tomo_obj, gene_ID, axes)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID (string)}

\item{axes}{axis (1, 2 or 3)}
}
\description{
Plot expression of single gene along an axis
}
\note{
You can do the same thing with
\code{tomo_obj$plot1dExpression(gene_ID, axes)}.
}
