% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{plot1dExpression}
\alias{plot1dExpression}
\title{plot expression of single gene along axes}
\usage{
plot1dExpression(tomo_obj, gene_ID, axes)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID (string)}

\item{axes}{axes (1, 2 or 3)}
}
\description{
plot expression of single gene along axes
}
