% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{PlotLossFunction}
\alias{PlotLossFunction}
\title{Plot the trend of the value of the loss function.}
\usage{
PlotLossFunction(tomoObj, geneID)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID (string)}
}
\description{
Plot the trend of the value of the loss function.
}
\note{
You can do the same things with
\code{tomoObj$PlotLossFunction(geneID)}.
}
