% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{plotLossFunction}
\alias{plotLossFunction}
\title{plots the trend of the value of the loss function.}
\usage{
plotLossFunction(tomo_obj, gene_ID)
}
\arguments{
\item{tomo_obj}{tomo_seq object}

\item{gene_ID}{single gene ID (string)}
}
\description{
plots the trend of the value of the loss function.
}
