% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_function.R
\name{ToDataFrame}
\alias{ToDataFrame}
\title{Convert reconstructed matrix to data.frame.}
\usage{
ToDataFrame(tomoObj, geneID)
}
\arguments{
\item{tomoObj}{tomoSeq object}

\item{geneID}{single gene ID}
}
\description{
Convert reconstructed matrix to data.frame.
}
