% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomoseqr.R
\name{tomo_seq}
\alias{tomo_seq}
\title{The class that contain hogefuga}
\description{
The class that contain hogefuga

The class that contain hogefuga
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{gene_list}}{hogefuga}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tomo_seq$new()}}
\item \href{#method-estimate3dExpressions}{\code{tomo_seq$estimate3dExpressions()}}
\item \href{#method-plotLossFunction}{\code{tomo_seq$plotLossFunction()}}
\item \href{#method-toDataFrame}{\code{tomo_seq$toDataFrame()}}
\item \href{#method-getReconstructedResult}{\code{tomo_seq$getReconstructedResult()}}
\item \href{#method-animate2d}{\code{tomo_seq$animate2d()}}
\item \href{#method-plot1dExpression}{\code{tomo_seq$plot1dExpression()}}
\item \href{#method-plot1dAllExpression}{\code{tomo_seq$plot1dAllExpression()}}
\item \href{#method-clone}{\code{tomo_seq$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make tomoSeq object. This method is not available to users. Instead of it,
\code{\link[=makeTomoObjSet]{makeTomoObjSet()}} is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$new(x, y, z, mask_shape = "rectangle")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A data.frame object containing a simulated Tomo-seq data for x-axis sections.
The rows represent genes. The first column contains gene IDs and the second and
subsequent columns contain gene expression levels in sections.}

\item{\code{y}}{A data.frame object containing a simulated Tomo-seq data for y-axis sections.
The rows represent genes. The first column contains gene IDs and the second and
subsequent columns contain gene expression levels in sections.}

\item{\code{z}}{A data.frame object containing a simulated Tomo-seq data for z-axis sections.
The rows represent genes. The first column contains gene IDs and the second and
subsequent columns contain gene expression levels in sections.}

\item{\code{mask_shape}}{Shape of mask. You can choose the value from \code{"rectangle"}, \code{"round"} or \code{"halfround"}. The default is \code{"rectangle"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate3dExpressions"></a>}}
\if{latex}{\out{\hypertarget{method-estimate3dExpressions}{}}}
\subsection{Method \code{estimate3dExpressions()}}{
Reconstructs 3D expression patterns of genes which are specified. See also \code{\link[=estimate3dExpressions]{estimate3dExpressions()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$estimate3dExpressions(queries = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queries}}{A vector consists of gene IDs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLossFunction"></a>}}
\if{latex}{\out{\hypertarget{method-plotLossFunction}{}}}
\subsection{Method \code{plotLossFunction()}}{
Plot the transition of the value of the loss function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$plotLossFunction(gene_ID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene_ID}}{a gene ID as string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-toDataFrame}{}}}
\subsection{Method \code{toDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$toDataFrame(gene_ID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getReconstructedResult"></a>}}
\if{latex}{\out{\hypertarget{method-getReconstructedResult}{}}}
\subsection{Method \code{getReconstructedResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$getReconstructedResult(gene_ID)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animate2d"></a>}}
\if{latex}{\out{\hypertarget{method-animate2d}{}}}
\subsection{Method \code{animate2d()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$animate2d(
  gene_ID,
  target,
  axes1,
  axes2,
  main,
  xlab,
  ylab,
  file,
  zlim,
  interval,
  aspect_ratio = c()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot1dExpression"></a>}}
\if{latex}{\out{\hypertarget{method-plot1dExpression}{}}}
\subsection{Method \code{plot1dExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$plot1dExpression(gene_ID, axes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot1dAllExpression"></a>}}
\if{latex}{\out{\hypertarget{method-plot1dAllExpression}{}}}
\subsection{Method \code{plot1dAllExpression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$plot1dAllExpression(axes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tomo_seq$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
